Moving the ThirdPartyScreen to the screens directory.
